# Fix quadratic construction time of `QuantumCircuit`